revert: "feat: uses generalizable chat completion adapter rather than one specific to dbrx-instruct"
This reverts commit f0f55d672fb48af72392a35eb234774e62e6ef6b.
